refactor: 우승자 판단 로직을 클래스로 분리
